+++ packages/create-streamdeck-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-streamdeck-react
 
+## 0.1.7
+
+### Patch Changes
+
+- update devtools, docs, bump packages
+
 ## 0.1.6
 
 ### Patch Changes